Put everything in place
