+++ templates/default/_editorconfig
