+++ templates/default/_gitignore
