+++ templates/default/_jshintrc
