+++ templates/default/gulp/tasks/dist.js
