+++ templates/default/gulp/util/handleErrors.js
