+++ templates/default/gulpfile.js
