+++ templates/default/package.json
