+++ templates/default/project.config.js
