+++ templates/default/project/scripts/global.js
