+++ templates/default/project/scripts/main.js
