+++ templates/default/project/scripts/prefabs/rotate-logo.js
