+++ templates/default/project/scripts/states/boot.js
