+++ templates/default/project/scripts/states/game.js
