+++ templates/default/project/scripts/utils/analytics.js
