+++ templates/default/static/app.manifest
