+++ templates/default/static/browserconfig.xml
