+++ templates/default/static/icon/apple-touch-icon-precomposed.png
