+++ templates/prefabs/sprite.js
